parse-tx/requirements.txt
requests
beautifulsoup4
pandas
scipy
scikit-learn
wheel
liac-arff